social_media_app/urls.py
from rest_framework import routers

from social_media_app.views import ProfileViewSet, PostViewSet

router = routers.DefaultRouter()
router.register("profiles", ProfileViewSet)
router.register("posts", PostViewSet)

urlpatterns = router.urls

app_name = "social"
