requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
black==23.7.0
charset-normalizer==3.2.0
click==8.1.5
colorama==0.4.6
Django==4.2.3
django-debug-toolbar==4.1.0
djangorestframework==3.14.0
drf-spectacular==0.26.3
frozenlist==1.4.0
idna==3.4
inflection==0.5.1
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.9.1
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.0
rpds-py==0.9.2
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
yarl==1.9.2
